Task 6: Do you want to eat long numbers? Huh?

+++ lab_03_06_01/main.c
@@ -15,7 +15,7 @@
 typedef struct
 {
     size_t len;
-    int nums[N];
+    long long nums[N];
 } array_t;
 
 
@@ -79,24 +79,24 @@ int read_by_spiral(matrix_t *matrix)
     while (s_row_index < e_row_index && s_column_index < e_column_index)
     {
         for (int i = s_column_index; i < e_column_index; ++i)
-            if (scanf("%d", (matrix->rows + s_row_index)->nums + i) != 1)
+            if (scanf("%lld", (matrix->rows + s_row_index)->nums + i) != 1)
                 return INPUT_ERROR;
         s_row_index++;
         for (int i = s_row_index; i < e_row_index; ++i)
-            if (scanf("%d", (matrix->rows+ i)->nums + (e_column_index - 1)) != 1)
+            if (scanf("%lld", (matrix->rows+ i)->nums + (e_column_index - 1)) != 1)
                 return INPUT_ERROR;
         e_column_index--;
         if (s_row_index < e_row_index)
         {
             for (int i = e_column_index-1; i >= s_column_index; --i)
-                if (scanf("%d", (matrix->rows + (e_row_index-1))->nums + i) != 1)
+                if (scanf("%lld", (matrix->rows + (e_row_index-1))->nums + i) != 1)
                     return INPUT_ERROR;
             e_row_index--;
         }
         if (s_column_index < e_column_index)
         {
             for (int i = e_row_index-1; i >= s_row_index; --i)
-                if (scanf("%d", (matrix->rows + i)->nums + s_column_index) != 1)
+                if (scanf("%lld", (matrix->rows + i)->nums + s_column_index) != 1)
                     return INPUT_ERROR;
             s_column_index++;
         }
